refactor: extract `RedirectCard` from `App`
